+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * improve TypeScript types for typing animation [#794](https://github.com/jcubic/jquery.terminal/issues/794)
 * add missing mousewheel typing animation [#795](https://github.com/jcubic/jquery.terminal/issues/795)
 * fix iterate formatting over closing bracket [#792](https://github.com/jcubic/jquery.terminal/issues/792)
+* fix resolving promise from echo typing animation when previous command was async [#797](https://github.com/jcubic/jquery.terminal/issues/797)
 
 ## 2.33.3
 ### Bugfix